requirements.txt
Flask
requests
gunicorn
pyyaml